fs-online-benchmark/locust-vertex/requirements.txt
locust
google-cloud-aiplatform
google-auth
pandas